Treat recent commits as a variant of unpushed commits
By default we insert either a log of commits that have not been pushed
to the upstream yet, or if there are no such commits, then a log of
recent commits.

Until now we used different section types and values for these
sections and instead used a kludge to treat those two sections as the
same section in some places.

The reason we did this was that we assumed that there would be places
where treating these sections as the same section would be wrong, and
that it would be better to explicitly handle them as the same section
in the places where we know that this should be done, instead of
undoing that in places where it shouldn't be done (and where we might
forget to do so).

As it turns out there is only one minor place where treating these
sections as the same section is a bit odd, while we failed to fully
implement the faking in some place where it would be needed.  The
section visibility was not always preserved and refreshing sometimes
jumped to a neighboring section.  There might have been other issues.

Now we use `unpushed' as the type and "@{upstream}.." as the value of
the "Recent commits" section, just like for the real "Unmerged into
<upstream>" section.

This obviously isn't correct, but currently the only downside is that,
"d d" on "Recent commits" shows the empty diff off the zero unpushed
commits.