Unset $GIT_{DIR,WORK_TREE} when loading magit
Setting these environment variables for an emacs instance is a recipe
for disaster.  The problem is that environment variables are shared
across all of Emacs, so every buffer shares the same values.

Unfortunately git commit does export $GIT_DIR before calling $EDITOR,
leading to issues such as #3516, so Magit now unsets that variable
and $GIT_WORK_TREE if they are set when `magit' is loaded.