Change order and initial visibility of logs in status buffer
The old order and default visibility of these sections was chosen to
emphasize the commits the user is less aware of.  If you work on a
feature branch, then you naturally know that you have created commits
that haven't made it into "master" yet.  So that section was hidden by
default and placed after some other log sections.  On the other hand
you might not be aware that someone else has pushed additional work to
the public version of the feature branch ("@{push}") or the upstream,
and those commits might be relevant to what you are currently doing.
So these sections were expanded by default and placed before other
log sections, maximizing the likelihood that you would notice.

In the past users have asked me to change this, but it took me until
now to get sufficiently annoyed by the behavior described above to
change it.  Now we maximize for usefulness.

The logs that are relevant more often than the others are placed first
and not hidden by default now.  It is still easy enough to notice that
there are new commits on relevant branches even when the respective
sections come later and are collapsed initially.  The logs of unpushed
commits should *normally* not get so long that they push the logs of
unpulled commits beyond the end of the window even when it is scrolled
to the top.